% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSI.R
\name{DSI.Guilds}
\alias{DSI.Guilds}
\title{Wrapper to calculate DSI for several different guilds}
\usage{
DSI.Guilds(Int, Phylo, Abund, Rep = 200)
}
\description{
Wrapper to calculate DSI for several different guilds
}

