% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSICom.R
\name{DSIpart}
\alias{DSIpart}
\title{Wrapper to calculate DSImean and the corresponding null model and provide proper results}
\usage{
DSIpart(Mat, rep = 999)
}
\description{
Wrapper to calculate DSImean and the corresponding null model and provide proper results
}
