% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManip.R
\name{MatToArray}
\alias{MatToArray}
\title{Organise a set of interaction matrices into one array}
\usage{
MatToArray(mats)
}
\arguments{
\item{mats}{A vector with matrix object names.}
}
\description{
Organises a set of local interaction matrices into one array for interactions
in all locations
}
