% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSICom.R
\name{NullNA}
\alias{NullNA}
\title{Null model to test for the different components calculated through DSImean}
\usage{
NullNA(DSILoc, rep = 999)
}
\description{
Null model to test for the different components calculated through DSImean
}
