% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSI.R
\name{dsi}
\alias{dsi}
\title{Distance-based specialisation index}
\usage{
dsi(Int, Dist, Abund, Rep = 999)
}
\arguments{
\item{Int}{Is an interaction array for which Specialization is to be measured. 
Consumers are in rows, resources in the columns and different locations 
(or other "local" partition - e.g. time) are slices. 
When the matrix has two dimensions no local variation is assumed.}

\item{Dist}{Object of class phylo with the $tip.label information matching Abundance data. 
The phylogeny should have all resource classes present in \code{Abund} 
and will be pruned to remove resource classes absent from \code{Abund}. 
Alternatively, a distance matrix, with dissimilarities between resource classes. 
Dimension names must match \code{Abund}.}

\item{Abund}{Matrix of local abundances/sampling effort of the resources in the 
interaction array at different locations. Resource items are in rows and locations in columns,
with labels as row and column names. Abundances can be inputted as relative or absolute values.}

\item{Rep}{Number of iterations for the null model, defaults to 999}
}
\value{
Returns an object of class \code{dsi}, with the following elements:
\item{data}{A list with the three objects used to calculate DSI - The interaction array, distance
matrix and resource availability matrix}
\item{consumers}{A character vector with the names of consumer species}
\item{resources}{A character vector with the names of resource classes (often species)}
\item{richness}{A numeric vector with the number of resource classes used by each consumer species}
\item{samp}{A numeric vector with the number of individuals or interactions recorded for each
 consumer species}
 \item{MPD}{A numeric vector with the abundance averaged mean pairwise distance 
 between resource items used by each consumer species}
 \item{null}{A matrix with \code{Rep} rows and \code{consumers} in columns, 
 with all \code{MPD} values obtained by the null model}
 \item{DSI}{A numeric vector with unstandardized DSI values. These are calculated as Z-scores 
 of the MPD values, compared to the null distribution of MPDs obtained from the null model. 
 It can be used to classify consumers as specialists, non-selective or generalists, 
 but is not appropriate for comparisons among species with different sampling effort}
 \item{class}{A character vector with the specialization class of consumers. 
 Singletons will return NA}
 \item{lim}{Theoretical maximum (for species with positive DSI) and minimum 
 (for species with negative DSI) DSI values attainable for the sample size and 
 resource similarity of each consumer species. Used to standardize DSI and generate comparable
 DSI* values. Maximum is obtained by assuming the species is a monophage, and minimum is
 calculated by using a Simulated annealing algorithm to distribute the recorded number
 of interactions among resource classes in such a manner that maximizes MPD}
 \item{DSIstar}{A numeric vector with standardized DSI values (DSI*). This is the index as 
 proposed in Jorge et al. (2017). DSI* values vary between -1 (extreme generalization) to 1 
 (extreme specialization). This should be the preferred value to be used when comparing species,
 as it has a very straightforward interpretation and is controlled for differences both in sampling 
 intensity, co-occurrence with resources and resource similarity differences.}
}
\description{
\code{dsi} measures the distance-based specialization indesx (DSI*) of a set of consumers. 
Relies on resource use information from an interaction matrix or array, 
resource similarities from a distance matrix or phylogeny, and local resource availability 
data for creating diet null models.
}
