% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSICom.R
\name{dsicom}
\alias{dsicom}
\title{Community level Distance-based specialisation index}
\usage{
dsicom(Int, Dist, Abund, Rep = 999, Part = TRUE)
}
\arguments{
\item{Int}{Is an interaction array for which Specialization is to be measured. 
Consumers are in rows, resources in the columns and different locations 
(or other "local" partition - e.g. time) are slices.}

\item{Dist}{Object of class phylo with the $tip.label information matching Abundance data. 
The phylogeny should have all resource species present in the Abundance data matrix 
and will be pruned accordingly. 
Alternatively, a distance matrix, with dissimilarities between resource classes. 
Dimension names must match Abundance data.}

\item{Abund}{Matrix of local abundances/sampling effort of the resources in the interaction array at different locations}

\item{Rep}{Number of iterations for the null model, defaults to 999}

\item{Part}{a logical indicating whether variability in DSI measured locally should be partitioned 
between species and localities.}
}
\description{
Measures the distance-based specialization indesx (DSI) of consumers in a set of communities. 
Relies on resource use information from an interaction matrix or array, 
resource similarities from a distance matrix or phylogeny, and local resource availability 
data standardizing the null models.
}
